refactor(modList): extract resolvePath into its own module
Move the pure path-resolution guard out of modList so it can be unit
tested without the Config / ipcWrapper dependencies.
